New working dir system (WIP)
