+++ test_files/CEASIOMpy_config_template/CEASIOMpy_config_template.cfg
@@ -1,0 +1,102 @@
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%          _____ ______           _____ _____ ____  __  __                     %
+%         / ____|  ____|   /\    / ____|_   _/ __ \|  \/  |                    %
+%        | |    | |__     /  \  | (___   | || |  | | \  / |_ __  _   _         %
+%        | |    |  __|   / /\ \  \___ \  | || |  | | |\/| | '_ \| | | |        %
+%        | |____| |____ / ____ \ ____) |_| || |__| | |  | | |_) | |_| |        %
+%         \_____|______/_/    \_\_____/|_____\____/|_|  |_| .__/ \__, |        %
+%                                                         |_|    |___/         %
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+% Configuration file                                                           %
+%                                                                              %
+% CEASIOMpy version: 1.0.0                                                     %
+% CPACS version: 3.3                                                           %
+% Case description: CEASIOMpy configuration file template                      %
+% Author: Aidan Jungo                                                          %
+% Date: 2022-01-06                                                             %
+%                                                                              %
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+% ---------------------------- INPUT/OUTPUT -----------------------------------%
+%
+% Abstolut or relative path to the input and output CPACS file.
+%
+WOKING_DIR = ./
+CPACS_TOOLINPUT = ./ToolInput.xml
+CPACS_TOOLOUTPUT = ./ToolOutput.xml
+%
+% ------------------------------ WORKFLOW -------------------------------------%
+%
+% Available modules
+%
+% Settings
+%    SettingsGUI : To open a GUI to set the option of the module after
+%
+% Geometry and mesh
+%    CPACSCreator :
+%    CPACS2SUMO :
+%    SUMOAutoMesh:
+%
+% Weight and balance
+%    WeightConventional
+%    WeightUnconventional
+%    BalanceConventional
+%    BalanceUnconventional
+%
+% Aerodynamics
+%    CLCalculator
+%    PyTornado
+%    SU2MeshDef
+%    SU2Run
+%    SkinFriction
+%    PlotAeroCoefficients
+%
+% Mission analysis
+%    Range
+%    StabilityStatic
+%
+% Surrogate modeling
+%    SMTrain
+%    SMUse
+%
+MODULE_PRE = (SettingsGUI, PyTornado)
+MODULE_OPTIM = (SettingsGUI,PyTornado)
+MODULE_POST = (PlotAeroCoefficients)
+%
+% ------------------------ OPTIMISATION OPTIONS -------------------------------%
+%
+% Optimisation methodes (OPTIM, DOE, NONE)
+%
+OPTIM_METHOD = NONE
+%
+% ----------------------------- SettingsGUI -----------------------------------%
+%
+% SetttingGUI parameters
+%
+%
+% ------------------------------- SU2Run --------------------------------------%
+%
+% SU2Run parameters
+%
+% Number of processor, if lager that the maximum number of proc, will use them all
+NB_PROC = 4
+%
+% Maximum number of iterations performed by SU2
+MAX_ITER = 500
+%
+% CFL Number, Courant–Friedrichs–Lewy condition (default value 1.0)
+CFL_Number = 1.0
+%
+% Path to exiting SU2 Mesh, if define, this one will be use
+SU2_MESH_PATH = ./MESH/mySU2mesh4.su2
+%
+% Extrage loads from the results
+EXTRACT_LOADS = TRUE
+%
+%
+% 
+% ------------------------------- CUSTOM --------------------------------------%
+% CUSTOM Parameter modification
+% Modification of other paramter in the CPACS file
+CUSTOM_XPATH_1 = /cpacs/vehicle/aircraft/...
+CUSTOM_VALUE_1 = 3
